requirements.txt
streamlit
soundfile
torch
transformers
openai-whisper
matplotlib
reportlab
